+++ hash.h
@@ -4,11 +4,11 @@
 #define M 5463458053
 
 struct file_hash {
-	int **list;
+	long long **list;
 	int count;
 };
 
-int integer_mod(int, long long);
+long long integer_mod(int, long long);
 struct file_hash* generate_hashes(const char*, int);
 char* substring(int start, int, char*);
 char** generate_k_grams(char*, int);
